feat(tslint): use extends instead of rulesDirectory
I misread the README and tried to use extends instead of rulesDirectory in `tslint.json` based on
other tslint config packages I use (tslint-config-prettier, tslint-microsoft-contrib, tslint-react).
This commit changes custom-tslint-rules' configuration to use rulesDirectory internally, which then
allows the user to add "custom-tslint-rules" to the extends property in `tslint.json`.

https://palantir.github.io/tslint/2016/03/31/sharable-configurations-rules.html
https://github.com/palantir/tslint-react/blob/3.0.0/tslint-react.json#L2
https://github.com/alexjoverm/tslint-config-prettier#get-started
https://github.com/Microsoft/tslint-microsoft-contrib/blob/b720cd9827a13c2878d304193544a5b030953ecb/tslint.json#L3

BREAKING CHANGE: this affects how users set up custom-tslint-rules, as they will need to change from
rulesDirectory to extends in their own `tslint.json` confiigurations (as shown in the README), so I
expect this to be a major breaking change (from 3.0.0 to 4.0.0).